Moved collection Inventory to PlayerInventoryData.cs
Changed recipe, deleted old ones
