--- Assets/Scripts/Score/ItemCollection.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: def1f0937341a4c9f89a333be034a0af
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
